requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0